FinalProject/generacion_0/generacion_1/generacion_2/MonitorAlerts/requirements.txt
Flask==2.0.2
pyserial==3.5